Basic OFDMA functions
"OFDMAPrepareMessage.m" receives a four character arrays, containing the
messages of four users, the number of total users, as well as the origin
and the destination nodes of the message. It adds the EOF and the extra
character to the end of the message and organizes them in a matrix.
"OFDMASignalGenerator.m" gets the output struct of
"OFDMAPrepareMessage.m" and makes it into an OFDMA signal. The script
tests "OFDMAPrepareMessage.m"